fix(nvim): set jdtls data directory only when using nvim-jdtls plugin
We can reliably build a `cmd` with a proper unique per-project data
directory only if we are in a ftplugin.
If we use native neovim lsp workflow, we can't change the `cmd` for every
project (hence no per-project unique data directory).

Previously it would work anyway because of the side effect of accessing
(and recomputing) the jdtls config from the ftplugin

Now, (if we don't use nvim-jdtls) no data directory is provided, so
jdtls will fallback to a temporary directory in /tmp; which should still
be better than having a single data directory for all projects like
setup by nvim-lspconfig